Moved blog page into its own app
