+++ content/blog/admin.py
@@ -2,4 +2,3 @@ from django.contrib import admin
 from content.blog import models
 
 admin.site.register(models.BlogIndexPage)
-admin.site.register(models.BlogPage)
