+++ content/blog_page/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from content.blog_page import models
+
+admin.site.register(models.BlogPage)
