+++ content/blog_page/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+from django.utils.translation import ugettext_lazy as _
+
+class BlogPageConfig(AppConfig):
+    name = 'content.blog_page'
+    verbose_name = _('BlogPage')
